Updated all .cs files to use file-scoped namespace declarations
Our internal code style has been to use file-scoped namespace declarations for a while now, which should therefore also be reflected in this package. Going forward, this will also be easier to compare changes between branches for newer versions of Umbraco, as we should also use file-scoped namespace declarations for those branches.

On the downside, it'll be a harder to compare changes between the U10 and U13 branches (unless we ignore whitespace). But we probably won't be making many changes to the U10 branch anyway.