Introduced factory method for back-office root nodes
The "RedirectsController.GetRootNodes" method is used for exposting the available root nodes to the back-office. Until now, the controller has handled convertering the "RedirectRootNode" instances received from "IRedirectsService" to their "RedirectRootNodeModel" counterparts used in the backoffice, but without any way to override this as the logic was placed in the controller.

With this commit, some of that logic is now delegated to the "RedirectsBackOfficeHelper.CreateRootNodeList" method for the overall list, and the "RedirectsBackOfficeHelper.CreateRootNode" method for the individual root nodes. Both methods are virtual, meaning that they can be extended in classes extending the "RedirectsBackOfficeHelper" class.